[alpha_factory] rename requests shim
