+++ alpha_factory_v1/demos/alpha_agi_business_v1/examples/openai_agent_client.py
@@ -15,7 +15,7 @@ from pathlib import Path
 try:
     import requests  # type: ignore
 except ModuleNotFoundError:  # pragma: no cover - offline shim
-    from alpha_factory_v1 import requests  # type: ignore
+    from alpha_factory_v1 import af_requests as requests  # type: ignore
 
 
 def _parse_args(argv: list[str] | None = None) -> argparse.Namespace:
