+++ alpha_factory_v1/demos/alpha_agi_business_v1/start_alpha_business.py
@@ -24,7 +24,7 @@ if ROOT not in sys.path:
 try:  # optional dependency
     import requests  # type: ignore
 except ModuleNotFoundError:  # pragma: no cover - offline shim
-    from alpha_factory_v1 import requests  # type: ignore
+    from alpha_factory_v1 import af_requests as requests  # type: ignore
 
 import check_env
